Add useDatasetPathDestination composable
Replaces compositeDatasetUtils and datasetPathDestination Vuex store.
